mvc: Realize MVC Resource:https://www.oracle.com/technical-resources/articles/javase/mvc.html
