+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/ChangeFieldModelListener.java
@@ -1,0 +1,8 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views;
+
+import java.util.EventListener;
+
+public interface ChangeFieldModelListener extends EventListener {
+
+    void eventChangeFiledModel(ChangeFieldModelEvent e);
+}
