--- src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/DialogFunc.java
@@ -1,11 +1,0 @@
-package bntu.fitr.gorbachev.ticketsgenerator.main.views;
-
-/**
- * This interface describes base functional {@link javax.swing.JDialog}
- */
-public interface DialogFunc {
-    /**
-     * This method initialization dialog window
-     */
-    void initDialog();
-}
