+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/gui/DialogFunc.java
@@ -1,0 +1,11 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.gui;
+
+/**
+ * This interface describes base functional {@link javax.swing.JDialog}
+ */
+public interface DialogFunc {
+    /**
+     * This method initialization dialog window
+     */
+    void initDialog();
+}
