+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/gui/panels/PanelType.java
@@ -1,0 +1,10 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.gui.panels;
+
+public enum PanelType {
+    ABOUT_PROGRAM,
+    ABOUT_AUTHOR,
+    MAIN_WINDOW,
+    SPLASH_SCREEN,
+    FILE_VIEWER,
+    RECORD_SETTING;
+}
