+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/model/impl/AboutAuthorModel.java
@@ -1,0 +1,8 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.model.impl;
+
+import bntu.fitr.gorbachev.ticketsgenerator.main.views.model.AbstractModel;
+
+
+public class AboutAuthorModel extends AbstractModel {
+
+}
