+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/model/impl/AboutProgramModel.java
@@ -1,0 +1,9 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.model.impl;
+
+import bntu.fitr.gorbachev.ticketsgenerator.main.views.model.AbstractModel;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class AboutProgramModel extends AbstractModel {
+}
