+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/model/impl/SplashScreenModel.java
@@ -1,0 +1,8 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.model.impl;
+
+import bntu.fitr.gorbachev.ticketsgenerator.main.views.model.AbstractModel;
+
+public class SplashScreenModel extends AbstractModel {
+
+
+}
