+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/model/impl/perconst/AboutAuthorFieldNameConst.java
@@ -1,0 +1,4 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.model.impl.perconst;
+
+public interface AboutAuthorFieldNameConst {
+}
