+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/model/impl/perconst/RecordSettingFieldNameConst.java
@@ -1,0 +1,6 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.model.impl.perconst;
+
+public interface RecordSettingFieldNameConst {
+    String QUANTITY_TICKET_ON_SINGLEPAGE = "QuantityTicketOnSinglePage";
+    String FONT_SIZE = "FontSize";
+}
