+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/model/impl/perconst/SplashScreenFiledNameConst.java
@@ -1,0 +1,5 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.views.model.impl.perconst;
+
+public interface SplashScreenFiledNameConst {
+
+}
