--- src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/views/panels/PanelType.java
@@ -1,10 +1,0 @@
-package bntu.fitr.gorbachev.ticketsgenerator.main.views.panels;
-
-public enum PanelType {
-    ABOUT_PROGRAM,
-    ABOUT_AUTHOR,
-    MAIN_WINDOW,
-    SPLASH_SCREEN,
-    FILE_VIEWER,
-    RECORD_SETTING;
-}
